Length recognition program for droplets
This folder contains a program that is able to detect and calculate the narrowest part of the droplets. Accuracy of this algorithm is not that high and there may be some tilted output images for this code, I am still working on how to solve them currently but it can be used as a sample in many conditions.